% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{collapse_newlines}
\alias{collapse_newlines}
\title{Replase single newline with a space, and ollapse multiple newlines into a single newline}
\usage{
collapse_newlines(text)
}
\arguments{
\item{text}{character string}
}
\value{
single character string
}
\description{
Replase single newline with a space, and ollapse multiple newlines into a single newline
}
