% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{collapse_whitespace}
\alias{collapse_whitespace}
\title{Collapse character vector into a single character string}
\usage{
collapse_whitespace(text)
}
\arguments{
\item{text}{character vector}
}
\value{
single character string
}
\description{
Collapse character vector into a single character string
}
