% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{cut_with_index}
\alias{cut_with_index}
\title{Cut a vector between the given indices (non-including the indices themselves)}
\usage{
cut_with_index(v, idxs)
}
\arguments{
\item{v}{character vector}

\item{idxs}{indices}
}
\value{
one character vector for each pair of indices
}
\description{
Cut a vector between the given indices (non-including the indices themselves)
}
