% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_doc}
\alias{get_doc}
\title{Get the documentation for a function in an R list}
\usage{
get_doc(function_name, package_name = NULL)
}
\arguments{
\item{function_name}{character name of function}

\item{package_name}{character name of package. If NULL, search within all
packages whose namespaces are loaded}
}
\value{
named list of documentation
}
\description{
Get the documentation for a function in an R list
}
