% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{get_rd_doc}
\alias{get_rd_doc}
\title{Get the raw Rd for the help for a function}
\usage{
get_rd_doc(function_name, package_name = NULL)
}
\arguments{
\item{function_name}{character name of function}

\item{package_name}{character name of package. If NULL, search within all
packages whose namespaces are loaded}
}
\value{
Rd document object
}
\description{
Get the raw Rd for the help for a function
}
