% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_tag}
\alias{get_tag}
\title{Get the \code{Rd_tag} of the node in an \code{Rd} document.}
\usage{
get_tag(rd_node)
}
\arguments{
\item{rd_node}{element in an Rd document}
}
\value{
characater string. 'NO TAG' if no tag was found
}
\description{
The \code{Rd_tag} gives meta-information about the Rd node and how to interpret it.
}
