% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse a section of an \code{Rd} object which has an \code{Rd_tag} of \code{arguments}}
\usage{
parse_arguments(rd_node)
}
\arguments{
\item{rd_node}{element in an Rd document}
}
\value{
named list
}
\description{
Parse a section of an \code{Rd} object which has an \code{Rd_tag} of \code{arguments}
}
