% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_describe}
\alias{parse_describe}
\title{Parse a block of text which has an \code{Rd_tag} of \code{describe}}
\usage{
parse_describe(rd_node)
}
\arguments{
\item{rd_node}{element in an Rd document}
}
\value{
named list
}
\description{
Parse a block of text which has an \code{Rd_tag} of \code{describe}
}
