% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_itemize}
\alias{parse_itemize}
\title{Parse an 'itemize' or 'enumerate'}
\usage{
parse_itemize(rd_node)
}
\arguments{
\item{rd_node}{element in an Rd document}
}
\value{
vector of character strings
}
\description{
Parse an 'itemize' or 'enumerate'
}
