% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_section}
\alias{parse_section}
\title{Parse a section from an Rd document.}
\usage{
parse_section(rd_section)
}
\arguments{
\item{rd_section}{An element from an \code{Rd} document which represents a
\verb{section} such as the \code{title} or \code{description}.}
}
\value{
contents of section
}
\description{
Parse a section from an Rd document given the document node for that section.
}
\details{
See \url{https://developer.r-project.org/parseRd.pdf} page 5 for a list
of all sectioning macros

Note: Only parsing \code{arguments}, \code{description}, \code{usage},
\code{title} and \code{name} for now.
}
