% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_text}
\alias{parse_text}
\title{Parse text from the given node in an \code{Rd} document}
\usage{
parse_text(rd_node)
}
\arguments{
\item{rd_node}{element in an Rd document}
}
\value{
character string
}
\description{
Parse text from the given node in an \code{Rd} document
}
