% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{rd2list}
\alias{rd2list}
\title{Parse documentation in Rd list format into a human-readable list}
\usage{
rd2list(rd_doc)
}
\arguments{
\item{rd_doc}{An \code{Rd} object as returned by \code{tools::parse_Rd} or
\code{get_rd_doc}.  An Rd doc is a list of lists with attributes.  The list
elements at the top-most level represent the sections within the documentation.}
}
\value{
structured character strings in a list object
}
\description{
This function is named to be similar to the \code{Rd2HTML} and \code{Rd2txt}
functions in the \pkg{tools} package.
}
\details{
In contrast to those functions, \code{Rd2list} can only be given an \code{Rd}
object.
}
